Refactored base class
